Refactor: Renamed the tiedote component on the frontpage
